requirements.txt
aiogram==3.15.0
openai==1.54.4
python-dotenv==1.0.0
chromadb>=1.5.1
PyPDF2==3.0.1
python-docx>=1.1.0
